Move isHadoopUnderFS calls to Hdfs factory
